# ENH, SIMD: Optimize the arithm fp implementation based on Highway Wra…