--- a/numpy/_core/meson.build
+++ b/numpy/_core/meson.build
@@ -922,13 +922,14 @@ umath_gen_headers = [
 foreach gen_mtargets : [
   [
     'loops_arithm_fp.dispatch.h',
-    src_file.process('src/umath/loops_arithm_fp.dispatch.c.src'),
+    'src/umath/loops_arithm_fp.dispatch.cpp',
     [
       [AVX2, FMA3], SSE2,
       ASIMD, NEON,
       VSX3, VSX2,
       VXE, VX,
       LSX,
+      RVV,
     ]
   ],
   [