--- a/numpy/_core/src/umath/loops_arithm_fp.dispatch.cpp
+++ b/numpy/_core/src/umath/loops_arithm_fp.dispatch.cpp
@@ -3,6 +3,7 @@
 #include "loops.h"
 #include "simd/simd.hpp"
 #include <hwy/highway.h>
+#include <limits>   // Required for std::numeric_limits
 
 namespace {
 using namespace np::simd;
@@ -625,6 +626,10 @@ complex_conjugate_square_func(char **args, npy_intp const *dimensions, npy_intp 
                 }
             }
             else if (sdst == 2 && static_cast<D>(ssrc) >= 0) {
+                if (hstep > 0 && ssrc > 0 && (std::numeric_limits<npy_intp>::max() / hstep < ssrc)) {
+                    // Overflow detected, fall back to the scalar loop
+                    goto loop_scalar;
+                }
                 auto i = Mul(hn::ShiftRight<1>(hn::Iota(_Tag<D>(), 0)), Set<D>(ssrc));
                 i = hn::OddEven(Add(i, Set<D>(1)), i);
                 for (; len >= vstep; len -= vstep, src += ssrc*vstep, dst += wstep) {