--- a/numpy/_core/src/umath/loops_arithm_fp.dispatch.cpp
+++ b/numpy/_core/src/umath/loops_arithm_fp.dispatch.cpp
@@ -625,7 +625,7 @@ complex_conjugate_square_func(char **args, npy_intp const *dimensions, npy_intp 
                     }
                 }
             }
-            else if (sdst == 2 && ssrc >= 0) {
+            else if (sdst == 2 && static_cast<D>(ssrc) >= 0) {
                 auto i = Mul(hn::ShiftRight<1>(hn::Iota(_Tag<D>(), 0)), Set<D>(ssrc));
                 i = hn::OddEven(Add(i, Set<D>(1)), i);
                 for (; len >= vstep; len -= vstep, src += ssrc*vstep, dst += wstep) {