Test environment: Banana Pi BPI-F3 (SpacemiT K1 8-core RISC-V chip) gcc version 14.2.0
Test command: spin bench --compare HEAD~1 -t bench_ufunc.ADD_SUB_MUL_DIV -t bench_ufunc.CONJUGATE_SQUARE_COMPLEX

Add class ADD_SUB_MUL_DIV and CONJUGATE_SQUARE_COMPLEX in benchmarks/benchmarks/bench_ufunc.py for this test on Banana Pi BPI-F3

```
class ADD_SUB_MUL_DIV(Benchmark):
    timeout = 10
    params = ['float32', 'float64', 'complex64', 'complex128']
    param_names = ['dtype']

    def setup(self, dtype):
        np.seterr(all='ignore')
        self.add_fn = np.add
        self.subtract_fn = np.subtract
        self.multiply_fn = np.multiply
        self.divide_fn = np.divide

        from .common import get_square_
        arr1 = get_square_(dtype)
        arr2 = get_square_(dtype).copy()

        self.data_add = (arr1, arr2)
        self.data_subtract = (arr1, arr2)
        self.data_multiply = (arr1, arr2)
        self.data_divide = (arr1, arr2)

    def time_add(self, dtype):
        self.add_fn(*self.data_add)

    def time_subtract(self, dtype):
        self.subtract_fn(*self.data_subtract)

    def time_multiply(self, dtype):
        self.multiply_fn(*self.data_multiply)

    def time_divide(self, dtype):
        self.divide_fn(*self.data_divide)


class CONJUGATE_SQUARE_COMPLEX(Benchmark):
    timeout = 10
    params = ['complex64', 'complex128']
    param_names = ['dtype']

    def setup(self, dtype):
        np.seterr(all='ignore')
        self.conjugate_fn = np.conjugate
        self.square_fn = np.square

        from .common import get_square_
        arr1 = get_square_(dtype)

        self.data_conjugate = (arr1,)
        self.data_square = (arr1,)

    def time_conjugate(self, dtype):
        self.conjugate_fn(*self.data_conjugate)

    def time_square(self, dtype):
        self.square_fn(*self.data_square)

```


(numpy_venv) root@yangwang:/home/yangwang/numpy# spin bench --compare HEAD~1 -t bench_ufunc.ADD_SUB_MUL_DIV -t bench_ufunc.CONJUGATE_SQUARE_COMPLEX
| Change   | Before [404b35ed] <main~1>   | After [ecab0fcc] <main>   |   Ratio | Benchmark (Parameter)                                          |
|----------|------------------------------|---------------------------|---------|----------------------------------------------------------------|
| -        | 165±4μs                      | 155±2μs                   |    0.94 | bench_ufunc.ADD_SUB_MUL_DIV.time_add('float64')                |
| -        | 247±5μs                      | 219±4μs                   |    0.89 | bench_ufunc.CONJUGATE_SQUARE_COMPLEX.time_square('complex128') |
| -        | 177±10μs                     | 157±4μs                   |    0.88 | bench_ufunc.ADD_SUB_MUL_DIV.time_multiply('float64')           |
| -        | 273±10μs                     | 233±4μs                   |    0.85 | bench_ufunc.ADD_SUB_MUL_DIV.time_add('complex128')             |
| -        | 311±20μs                     | 259±7μs                   |    0.83 | bench_ufunc.ADD_SUB_MUL_DIV.time_subtract('complex128')        |
| -        | 206±4μs                      | 169±4μs                   |    0.82 | bench_ufunc.ADD_SUB_MUL_DIV.time_subtract('float64')           |
| -        | 123±6μs                      | 99.0±7μs                  |    0.8  | bench_ufunc.ADD_SUB_MUL_DIV.time_add('float32')                |
| -        | 244±20μs                     | 195±4μs                   |    0.8  | bench_ufunc.ADD_SUB_MUL_DIV.time_multiply('complex64')         |
| -        | 243±2μs                      | 177±8μs                   |    0.73 | bench_ufunc.ADD_SUB_MUL_DIV.time_subtract('complex64')         |
| -        | 226±8μs                      | 163±4μs                   |    0.72 | bench_ufunc.ADD_SUB_MUL_DIV.time_add('complex64')              |
| -        | 138±5μs                      | 98.2±7μs                  |    0.71 | bench_ufunc.ADD_SUB_MUL_DIV.time_subtract('float32')           |
| -        | 226±8μs                      | 153±9μs                   |    0.68 | bench_ufunc.CONJUGATE_SQUARE_COMPLEX.time_square('complex64')  |
| -        | 378±20μs                     | 235±9μs                   |    0.62 | bench_ufunc.ADD_SUB_MUL_DIV.time_divide('float64')             |
| -        | 241±0.6μs                    | 139±4μs                   |    0.58 | bench_ufunc.ADD_SUB_MUL_DIV.time_divide('float32')             |